server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pymongo==4.8.0
sentence-transformers==3.1.1
scikit-learn==1.5.1
numpy==1.26.4
python-dotenv==1.0.1
langchain
langchain-core
langchain-community
langchain-groq
requests
slowapi